+++ server/src/main/java/com/moovda_project/moovda/module/comment/dto/CommentPostDto.java
@@ -9,6 +9,5 @@ import javax.validation.constraints.Min;
 @Getter
 public class CommentPostDto {
     private String content;
-
     private double star;
 }
