+++ server/src/main/java/com/moovda_project/moovda/module/comment/dto/CommentResponseDto.java
@@ -1,8 +1,9 @@
-package com.moovda_project.moovda.module.movie.dto;
+package com.moovda_project.moovda.module.comment.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Setter
@@ -11,10 +12,11 @@ public class CommentResponseDto {
 //    String nickname;
 
     Long memberId;
+
     Long commentId;
     String content;
     Double star;
     Integer likeCount;
-
+    LocalDateTime createdAt;
 
 }
