+++ server/src/main/java/com/moovda_project/moovda/module/like/controller/LikeController.java
@@ -1,7 +1,7 @@
-package com.moovda_project.moovda.module.comment.controller;
+package com.moovda_project.moovda.module.like.controller;
 
 import com.moovda_project.moovda.global.utils.MemberIdExtractor;
-import com.moovda_project.moovda.module.comment.service.LikeService;
+import com.moovda_project.moovda.module.like.service.LikeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
