+++ server/src/main/java/com/moovda_project/moovda/module/movie/dto/MovieFilterResponseDto.java
@@ -12,5 +12,6 @@ public class MovieFilterResponseDto {
     long movieId;
     String title;
     String poster;
+    Double starAvg;
 
 }
