+++ server/src/main/java/com/moovda_project/moovda/module/movie/dto/genre/GenreResponseDto.java
@@ -1,4 +1,4 @@
-package com.moovda_project.moovda.module.movie.dto;
+package com.moovda_project.moovda.module.movie.dto.genre;
 
 import lombok.Getter;
 import lombok.Setter;
