+++ server/src/main/java/com/moovda_project/moovda/module/movie/dto/staff/StaffResponseDto.java
@@ -1,4 +1,4 @@
-package com.moovda_project.moovda.module.movie.dto;
+package com.moovda_project.moovda.module.movie.dto.staff;
 
 import lombok.Getter;
 import lombok.Setter;
