+++ server/src/main/java/com/moovda_project/moovda/module/watch/entity/Watched.java
@@ -1,4 +1,4 @@
-package com.moovda_project.moovda.module.movie.entity.watch;
+package com.moovda_project.moovda.module.watch.entity;
 
 import com.moovda_project.moovda.global.audit.Auditable;
 import com.moovda_project.moovda.module.member.entity.Member;
